[#780] Move cpu set to common
